Fixed Case Sensitive Issue with Satellite name
1) set names with git mv to lower_case
