+++ LDAR_Sim/model_code/satellite_company.py
@@ -19,10 +19,10 @@
 #
 # ------------------------------------------------------------------------------
 
+
 import numpy as np
-from satellite import satellite
-from generic_functions import get_prop_rate
-import math
+from orbit_predictor.sources import get_predictor_from_tle_lines
+import netCDF4 as nc
 
 
 class satellite_company:
@@ -44,6 +44,38 @@ class satellite_company:
         self.timeseries['satellite_flags_redund2'] = np.zeros(self.parameters['timesteps'])
         self.timeseries['satellite_flags_redund3'] = np.zeros(self.parameters['timesteps'])
         self.timeseries['satellite_sites_visited'] = np.zeros(self.parameters['timesteps'])
+		
+		
+		# load pre-defined orbit grids 
+        Dataset = nc.Dataset(self.parameters['methods']['Satellite']['Satellite_grid'],'r')
+        self.satgrid = Dataset.variables['sat'][:]
+        Dataset.close()
+        
+        # load cloud cover data
+        Dataset = nc.Dataset(self.parameters['methods']['Satellite']['CloudCover'],'r')
+        self.cloudcover = Dataset.variables['tcc'][:]
+        Dataset.close()
+        
+        
+        # build a satellite orbit object 
+        sat = self.parameters['methods']['Satellite']['name']
+        tlefile = self.parameters['methods']['Satellite']['TLE_file']
+        TLEs = []  
+        with open(tlefile) as f:
+            for line in f:
+                TLEs.append(line.rstrip())   
+
+        i = 0 
+        for x in TLEs: 
+            if x == sat: 
+                break
+            i+=1
+        TLE_LINES = (TLEs[i+1],TLEs[i+2])
+        
+        self.predictor = get_predictor_from_tle_lines(TLE_LINES)
+        
+        return
+    
 
         # Assign the correct follow-up threshold
         if self.config['follow_up_thresh'][1] == "absolute":
@@ -66,21 +98,22 @@ class satellite_company:
         # Initialize 2D matrices to store deployment day (DD) counts and MCBs
         self.DD_map = np.zeros(
             (len(self.state['weather'].longitude),
-             len(self.state['weather'].latitude)))
+              len(self.state['weather'].latitude)))
         self.MCB_map = np.zeros(
             (len(self.state['weather'].longitude),
-             len(self.state['weather'].latitude)))
+              len(self.state['weather'].latitude)))
 
-        # Initialize the individual satellite (the agents)
+         # Initialize the individual aircraft crews (the agents)
         for i in range(config['n_crews']):
-            self.crews.append(satellite(state, parameters, config,
-                                        timeseries, self.deployment_days, id=i + 1))
+            self.crews.append(Satellite(state, parameters, config,
+                                             timeseries, self.deployment_days, id=i + 1))
 
         return
 
+
     def find_leaks(self):
         """
-        The satellite company tells satellites to find leaks.
+        The satellite company tells all the crews to get to work.
         """
 
         self.candidate_flags = []
@@ -107,69 +140,11 @@ class satellite_company:
                                     site['lat_index'],
                                     self.state['t'].current_timestep]:
                 available_sites += 1
-        if available_sites == 0:
-            prop_avail = 0
-        else:
-            prop_avail = available_sites / len(self.state['sites'])
-        self.timeseries['satellite_prop_sites_avail'].append(prop_avail)
+        prop_avail = available_sites / len(self.state['sites'])
+        self.timeseries['aircraft_prop_sites_avail'].append(prop_avail)
 
         return
 
-    def flag_sites(self):
-        """
-        Flag the most important sites for follow-up.
-        """
-        # First, figure out how many sites you're going to choose
-        n_sites_to_flag = len(self.candidate_flags) * self.config['follow_up_ratio']
-        n_sites_to_flag = int(math.ceil(n_sites_to_flag))
-
-        sites_to_flag = []
-        site_cum_rates = []
-
-        for i in self.candidate_flags:
-            site_cum_rates.append(i['site_cum_rate'])
-
-        site_cum_rates.sort(reverse=True)
-        target_rates = site_cum_rates[:n_sites_to_flag]
-
-        for i in self.candidate_flags:
-            if i['site_cum_rate'] in target_rates:
-                sites_to_flag.append(i)
-
-        for i in sites_to_flag:
-            site = i['site']
-            leaks_present = i['leaks_present']
-            site_cum_rate = i['site_cum_rate']
-            venting = i['venting']
-
-            # If the site is already flagged, your flag is redundant
-            if site['currently_flagged']:
-                self.timeseries['satellite_flags_redund1'][self.state['t'].current_timestep] += 1
-
-            elif not site['currently_flagged']:
-                # Flag the site for follow up
-                site['currently_flagged'] = True
-                site['date_flagged'] = self.state['t'].current_date
-                site['flagged_by'] = self.config['name']
-                self.timeseries['satellite_eff_flags'][self.state['t'].current_timestep] += 1
-
-                # Does the chosen site already have tagged leaks?
-                redund2 = False
-                for leak in leaks_present:
-                    if leak['date_tagged'] is not None:
-                        redund2 = True
-
-                if redund2:
-                    self.timeseries['satellite_flags_redund2'][
-                        self.state['t'].current_timestep] += 1
-
-                # Would the site have been chosen without venting?
-                if self.parameters['consider_venting']:
-                    if (site_cum_rate - venting) < self.config['follow_up_thresh']:
-                        self.timeseries['satellite_flags_redund3'][
-                            self.state['t'].current_timestep] += 1
-
-            return
 
     def site_reports(self):
         """
@@ -180,3 +155,5 @@ class satellite_company:
         for site in self.state['sites']:
             site['satellite_prop_DDs'] = self.DD_map[site['lon_index'], site['lat_index']]
             site['satellite_MCB'] = self.MCB_map[site['lon_index'], site['lat_index']]
+
+        return
